+++ src/main/java/com/system/billingSystem/service/CustomerService.java
@@ -1,7 +1,6 @@
 package com.system.billingSystem.service;
 
 import com.system.billingSystem.dto.CustomerDto;
-import com.system.billingSystem.exeption.ResourceNotFoundException;
 import com.system.billingSystem.model.Customer;
 import com.system.billingSystem.repository.CustomerRepository;
 import jakarta.transaction.Transactional;
@@ -36,7 +35,8 @@ public class CustomerService {
     public CustomerDto delete(Long id) {
         try{
             CustomerDto customerDto = this.findById(id);
-            this.customerRepository.deleteById(id);
+            if (customerDto != null)
+                this.customerRepository.deleteById(id);
             return customerDto;
         }catch (Exception e){
             logger.log(Level.SEVERE, "Error in CustomerService on method delete id: " + id);
@@ -47,12 +47,9 @@ public class CustomerService {
     public CustomerDto findById(Long id) {
         try {
             Customer customer = this.customerRepository.findById(id).orElse(null);
-
             if (customer != null)
                 return new CustomerDto(customer);
-            else
-                throw new ResourceNotFoundException("Customer not found for this id :: " + id);
-
+            return null;
         }catch (Exception e){
             logger.log(Level.SEVERE, "Error in CustomerService on method findById on id: " + id);
             throw e;
